Getting Create and Delete Works

+++ src/models/works/works.service.ts
@@ -1,5 +1,6 @@
+import { UserEntity } from '#models/users/serializers/users.serializer';
 import { WorksRepository } from './works.repository';
-import { ConflictException, Injectable } from '@nestjs/common';
+import { Injectable, HttpException, HttpStatus } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { WorkEntity } from './serializers/works.serializer';
 import { Work } from './entities/works.entity';
@@ -12,18 +13,35 @@ export class WorksService {
     private readonly worksRepository: WorksRepository,
   ) {}
 
-  async create(createWorkDto: CreateWorkDto): Promise<WorkEntity> {
-    return await this.worksRepository.createEntity(createWorkDto, ['icon']);
+  async create(
+    createWorkDto: CreateWorkDto,
+    user: UserEntity,
+  ): Promise<WorkEntity> {
+    return await this.worksRepository.createEntity({ user, ...createWorkDto }, [
+      'user',
+      'platform',
+      'thumbnail',
+      'tools',
+    ]);
   }
 
   async getAll(): Promise<Work[]> {
-    return await this.worksRepository.getAll(['icon']);
+    return await this.worksRepository.getAll([
+      'user',
+      'platform',
+      'thumbnail',
+      'tools',
+    ]);
   }
 
-  async delete(id: string): Promise<boolean> {
+  async delete(id: string, user: UserEntity): Promise<boolean> {
+    const work = await this.worksRepository.get(id, ['user'], true);
+    if (work.user.id !== user.id) {
+      throw new HttpException(`it's not your post.`, HttpStatus.FORBIDDEN);
+    }
     const { affected } = await this.worksRepository.delete({ id });
     if (affected < 1) {
-      throw new ConflictException();
+      throw new HttpException('operation failed.', HttpStatus.CONFLICT);
     }
     return true;
   }
